backend/requirements.txt
asgiref==3.8.1
boto3==1.39.1
botocore==1.39.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
razorpay==1.4.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.9.0
